Tktable v2.2
Undid caching of activeLayout in favor of activeTagPtr because I was
abusing the nature of the token returned by Tk_ComputeTextLayout,
causing amazingly odd cores.

Changed TableGetLastCell definition to fix scrollbar bug. (Brutti)

Changed "#if (TK_MINOR_VERSION > 0)" to "#ifdef TCL_UTF_MAX" to
remove the ambiguity of purpose (for UTF code).

Fixed bug in TableInsertChars that mangled the purpose of -autoclear,
as well as confusing icursor index when using it.