Bug fix for new tags using the option database for default values instead of the default value.
When porting to use Tk object interface for Tk_ConfigureWidget, the TK_CONFIG_ARGV_ONLY flag was removed.
This changed the behavior so if a option is not specified in the command args, then new tags would use the option database value if a match is found, instead of the default value. The fix is the add back the TK_CONFIG_ARGV_ONLY flag.

Bug fix for https://github.com/bohagan1/TkTable/issues/3